src/requirements.txt
django==4.0.3
djangorestframework

gunicorn==20.1.0

psycopg2-binary==2.9.3